Move system header inclusions outside extern "C"
GLib and GTK both include extern "C" guards in their header files, so it
is unnecessary to wrap them in an extern "C" block.

A redundant extern "C" block can be harmful, because header files might
test #ifdef __cplusplus and include C++ syntax if it is defined, even in
an otherwise C header. In particular, GLib has done this since version
2.68, in order to make macros like g_object_ref() type-safe when called
from C++ code. the result was that this project failed to compile.

Resolves: https://github.com/brummer10/gxtuner/issues/19
Bug-Debian: https://bugs.debian.org/992246