build: Target a specific GLib API version
Setting GLIB_VERSION_MIN_REQUIRED selects the minimum required version
of GLib for this project. Code that was deprecated after that version
will not cause deprecation warnings, and where header files have changed
their compile-time behaviour over time, the behaviour that was seen in
the selected version will be used where possible.

Setting GLIB_VERSION_MAX_ALLOWED causes GLib to emit warnings if a
function introduced after the selected version is used.

In particular, this disables new C++ behaviour introduced in GLib 2.68,
which caused this project to fail to build.

GLib 2.54 is an arbitrary choice: it happens to be the version that was
the current stable release when gxtuner 3.0.0 was released. It could
be reduced if gxtuner needs to compile on older distributions, or
increased if a newer dependency is acceptable.

Resolves: https://github.com/brummer10/gxtuner/issues/19
Bug-Debian: https://bugs.debian.org/992246